moved commands.py to cli
